+++ client/src/components/Forms/CategoryForm.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { Form, Modal, Button, } from "semantic-ui-react";
+import { Form, Modal, Button, Icon, } from "semantic-ui-react";
 import Dropzone from "react-dropzone";
 import axios from "axios";
 import { Image, CloudinaryContext } from "cloudinary-react";
@@ -97,7 +97,7 @@ class CategoryForm extends Component {
   render() {
     return (
       <>
-        <Modal trigger={<Button as="td">Edit</Button>}>
+        <Modal trigger={<div style={{}} as="td"><h2><Icon name='edit' /></h2></div>}>
           {this.categoryFormat()}
         </Modal>
       </>
